requirementss.txt
aistudio-sdk
annotated-types
anyio
bce-python-sdk
certifi
chardet
charset-normalizer
click
colorama
colorlog
contourpy
customtkinter
cvzone
cycler
darkdetect
easyocr
filelock
filterpy
fonttools
fsspec
future
h11
hf-xet
httpcore
httpx
huggingface_hub
idna
ImageIO
imagesize
Jinja2
kiwisolver
lap
lazy_loader
MarkupSafe
matplotlib
modelscope
mpmath
networkx
ninja
numpy
opencv-contrib-python
opencv-python
opt-einsum
packaging
paddleocr
paddlepaddle
paddlex
pandas
pillow
polars
polars-runtime-32
prettytable
protobuf
psutil
py-cpuinfo
pyclipper
pycryptodome
pydantic
pydantic_core
pyparsing
pypdfium2
pytesseract
python-bidi
python-dateutil
python-dotenv
pytz
PyYAML
requests
ruamel.yaml
ruamel.yaml.clib
safetensors
scikit-image
scipy
setuptools
shapely
shellingham
six
supabase
sympy
tifffile
torch
torchaudio
torchvision
tqdm
typer-slim
typing-inspection
typing_extensions
tzdata
ujson
ultralytics
ultralytics-thop
urllib3
wcwidth